update 2 wintools

+++ ATmega32U4/Functions.h
@@ -28,7 +28,7 @@
 #define WS2812_MASK		(1<<1)
 #define WS2812_SAVE		1			/*Power saver, divide light level with this.*/
 #define MAX_DELAY 0x0010
-#elif defined(LWL84)
+#elif defined(Vem84)
 #define WS2812_PORT		PORTC
 #define WS2812_DDR		DDRC
 #define WS2812_MASK		(1<<7)
@@ -36,7 +36,7 @@
 #define MAX_DELAY 0x0010
 #endif
 uint16_t rgb_rainbow[WS2812_COUNT];
-uint16_t rgb_pos[WS2812_COUNT];
+uint8_t rgb_pos[WS2812_COUNT];
 uint8_t rgb_fixcolor[(WS2812_COUNT * 3)];
 uint8_t rgb_type;
 uint8_t rgb_state;
